[REF] web: convert core>utils>functions tests to hoot task 3705027
